requirements.txt
streamlit==1.45.0
groq==0.25.0
plotly==6.0.1
pandas==2.2.3
numpy==2.2.6
python-dotenv==1.1.0
pydantic==2.11.7